Fixed weird bug where the questiongenerator loop loops through vars that were not in its keys, fixed it by using an extra check

+++ Jordy_Dennis/GUI/question_generator.py
@@ -26,34 +26,33 @@ class Question_Generator:
         toBeDeleteQuestions = self.questions.copy()
         if self.form:
             # for every question that is evaluated
-            for varName in self.questions:
-                # get information of the question
-                label = self.questions[varName].getQuestion()
-                var_type = self.varDict[varName]['node'].checkTypes()
-                value = self.varDict[varName]['node'].evaluate()
-                # check if assignment node, only show evaluated value
-                if (type(self.questions[varName]) == AssignmentNode):
-                    if (self.form.getQuestionFromPage(varName)):
-                        self.form.getQuestionFromPage(varName).setValue(value)
+            for varName in list(self.questions.keys()):
+                if varName in self.questions.keys():
+                    # get information of the question
+                    label = self.questions[varName].getQuestion()
+                    var_type = self.varDict[varName]['node'].checkTypes()
+                    value = self.varDict[varName]['node'].evaluate()
+                    # check if assignment node, only show evaluated value
+                    if (type(self.questions[varName]) == AssignmentNode):
+                        if (self.form.getQuestionFromPage(varName)):
+                            self.form.getQuestionFromPage(varName).setValue(value)
 
-                # if the question is not yet in the GUI
-                if (not self.form.isQuestionOnPage(varName)):
+                    # if the question is not yet in the GUI
+                    if (not self.form.isQuestionOnPage(varName)):
 
-                    # it is not the initial setup process
-                    if not initial:
-                        # delete every question that is under the to be inserted if-question
-                        for varNameToBeDeleted in toBeDeleteQuestions:
-                            self.form.removeQuestionFromPage(varNameToBeDeleted)
+                        # it is not the initial setup process
+                        if not initial:
+                            # delete every question that is under the to be inserted if-question
+                            for varNameToBeDeleted in toBeDeleteQuestions:
+                                self.form.removeQuestionFromPage(varNameToBeDeleted)
 
-                    # insert new question into the GUI
-                    self.form.addQuestionToPage(varName, label, var_type, value)
-                # delete question from the to be deleted list
-                del toBeDeleteQuestions[varName]
+                        # insert new question into the GUI
+                        self.form.addQuestionToPage(varName, label, var_type, value)
+                    # delete question from the to be deleted list
+                    del toBeDeleteQuestions[varName]
 
-                # remove if question if no longer valid
-                self.form.deleteInvalidQuestions(self.questions)
-
-        return self.questions
+                    # remove if question if no longer valid
+                    self.form.deleteInvalidQuestions(self.questions)
 
     # Create the list of all the questions by recursively looping through the statements and adding them to te dictionairy
     def get_questions(self, block):
